Moved list DividerItemDecoration class to a proper class
